LL4/demo2-deployment/requirements.txt
crewai>=0.80.0
flask>=2.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

